MRG: fix deploy by specifying main appfile, move to gh-actions only for deploy; add dev deploy for PRs  (#37)
Here, I:
1. **SPECIFY `appPrimaryDoc = "RWdbApp.R"` in the `rsconnect::deployApp`
call**. This is required because our app is not called `app.R` or `ui.R`
+ `server.R`
2. REMOVED .Rprofile files  -- these should not be used/included

This sorts deployment, so we see the actual app and not the prior page
error: "An error has occurred. Check your logs or contact the app author
for clarification."

3. Modified the github-actions workflow to avoid using docker for shiny
app deployment -- simpler, and allows package installations to be
cached, speeding up frequent workflow runs. To the basic setup, I added
all the libraries we need for geolocation, etc that we were previously
installing via the shiny deployment dockerfile.
4. Added a similar gh action for a development server that deploys for
Pull Requests only (and added its associated tokens/secrets). The dev
server uses my account, so it will deploy to:
https://bluegenes.shinyapps.io/RalstoniaWiltDashboard/. This allows
testing without breaking the main deployment.